requirements.txt
langchain==0.1.12
langchain-community>=0.0.28
langchain-groq>=0.0.3
chromadb==0.4.24
sentence-transformers==2.5.1
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
html5lib>=1.1
requests==2.31.0
playwright==1.42.0
pymupdf==1.23.26
python-docx==1.1.0
pytesseract==0.3.10
easyocr==1.7.1
sqlalchemy==2.0.28
transformers==4.38.2
torch==2.2.1
pillow==10.2.0
python-multipart==0.0.9
unstructured==0.12.5
markdown==3.5.2
pydantic==2.6.3
jinja2==3.1.3
aiohttp>=3.9.0 